+++ ts/components/Scrubber.tsx
@@ -1,12 +1,13 @@
 import * as React from "react";
-import {StreamScrubber, KeybindingProps} from "../interfaces";
+import {StreamScrubber, KeybindingProps, LocaleProps} from "../interfaces";
 import Timeline from "./Timeline";
 import SpeedSelector from "./SpeedSelector";
 import Tooltipper from "./Tooltipper";
 import {cookie} from "cookie_js";
 import * as _ from "lodash";
+import Settings from "./Settings";
 
-interface ScrubberProps extends KeybindingProps, React.ClassAttributes<Scrubber> {
+interface ScrubberProps extends KeybindingProps, LocaleProps, React.ClassAttributes<Scrubber> {
 	scrubber: StreamScrubber;
 	swapPlayers?: () => void;
 	isSwapped?: boolean;
@@ -21,6 +22,7 @@ interface ScrubberProps extends KeybindingProps, React.ClassAttributes<Scrubber>
 	onClickRevealCards?: () => void;
 	isLogVisible?: boolean;
 	toggleLog?: () => void;
+	onSelectLocale?: (locale: string, loaded?: () => void) => void;
 }
 
 interface ScrubberState {
@@ -31,6 +33,7 @@ interface ScrubberState {
 	speed?: number;
 	duration?: number;
 	currentTime?: number;
+	isShowingSettings?: boolean;
 }
 
 export default class Scrubber extends React.Component<ScrubberProps, ScrubberState> {
@@ -42,7 +45,7 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 
 		// restore speed setting
 		let speed = +cookie.get("joust_speed", "1");
-		if(Scrubber.SPEEDS.indexOf(speed) === -1) {
+		if (Scrubber.SPEEDS.indexOf(speed) === -1) {
 			speed = 1;
 		}
 
@@ -52,6 +55,7 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 			canRewind: false,
 			canPlay: false,
 			speed: speed,
+			isShowingSettings: false,
 		};
 		this.props.scrubber.setSpeed(this.state.speed);
 		this.onKeyDown = this.onKeyDown.bind(this);
@@ -62,13 +66,13 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 	}
 
 	public componentWillUpdate(nextProps: ScrubberProps, nextState: ScrubberState): void {
-		if(!_.isEqual(nextProps, this.props)) {
+		if (!_.isEqual(nextProps, this.props)) {
 			this.removeListeners(this.props);
 		}
 	}
 
 	public componentDidUpdate(prevProps: ScrubberProps, prevState: ScrubberState, prevContext: any): void {
-		if(!_.isEqual(prevProps, this.props)) {
+		if (!_.isEqual(prevProps, this.props)) {
 			this.registerListeners(this.props);
 		}
 	}
@@ -81,7 +85,9 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 			nextProps.fullscreenError === this.props.fullscreenError &&
 			nextProps.isRevealingCards === this.props.isRevealingCards &&
 			nextProps.canRevealCards === this.props.canRevealCards &&
-			nextProps.isLogVisible === this.props.isLogVisible) {
+			nextProps.isLogVisible === this.props.isLogVisible &&
+			nextProps.locale === this.props.locale
+		) {
 			return false;
 		}
 		return true;
@@ -108,7 +114,7 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 			return;
 		}
 
-		if(e.shiftKey || e.altKey || e.ctrlKey || e.metaKey) {
+		if (e.shiftKey || e.altKey || e.ctrlKey || e.metaKey) {
 			// do not trigger with modifier
 			return;
 		}
@@ -219,65 +225,104 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 	public render(): JSX.Element {
 		const playpause = this.state.playing ?
 			<Tooltipper title="Pause" align="left">
-				<button onClick={() => this.pause()} disabled={!this.state.canInteract} className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-pause"></i></button>
+				<button onClick={() => this.pause()} disabled={!this.state.canInteract}
+						className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-pause"></i></button>
 			</Tooltipper> :
 			<Tooltipper title="Play" align="left">
-				<button onClick={() => this.play()} disabled={!this.state.canPlay} className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-play"></i></button>
+				<button onClick={() => this.play()} disabled={!this.state.canPlay}
+						className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-play"></i></button>
 			</Tooltipper>;
 
 		const restart = <Tooltipper title="Restart" align="left">
-			<button onClick={() => this.rewind()} disabled={!this.state.canRewind} className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-fast-backward"></i></button>
+			<button onClick={() => this.rewind()} disabled={!this.state.canRewind}
+					className="joust-scrubber-button-wide"><i className="joust-fa joust-fa-fast-backward"></i></button>
 		</Tooltipper>;
 
-		const fullscreen = this.props.isFullscreen ?
-			<Tooltipper title="Minimize" align="right">
-				<button onClick={this.props.onClickMinimize}><i className="joust-fa joust-fa-compress"></i></button>
+		const speedSelector = <Tooltipper title="Playback speed" desktop="%s (W/S)">
+			<SpeedSelector
+				speed={this.state.speed}
+				speeds={Scrubber.SPEEDS}
+				selectSpeed={(speed: number) => this.selectSpeed(speed)}
+				disabled={!this.state.canInteract}
+			/>
+		</Tooltipper>;
+
+		const fullscreenToggle = this.props.isFullscreen ?
+			<Tooltipper title="Restore" align="right" desktop="%s (F)">
+				<button onClick={this.props.onClickMinimize}><i className="joust-fa joust-fa-window-restore"></i>
+				</button>
 			</Tooltipper> :
-			<Tooltipper title={this.props.fullscreenError ? "Error entering fullscreen." : "Fullscreen"} align="right" forceShow={this.props.fullscreenError}>
-				<button onClick={() => this.props.isFullscreenAvailable && this.props.onClickFullscreen()} disabled={!this.props.isFullscreenAvailable}><i className="joust-fa joust-fa-expand"></i></button>
+			<Tooltipper
+				title={this.props.fullscreenError ? "Error entering fullscreen." : "Fullscreen"}
+				align="right"
+				forceShow={this.props.fullscreenError} desktop="%s (F)"
+			>
+				<button
+					onClick={() => this.props.isFullscreenAvailable && this.props.onClickFullscreen()}
+					disabled={!this.props.isFullscreenAvailable}
+				>
+					<i className="joust-fa joust-fa-arrows-alt"></i>
+				</button>
 			</Tooltipper>;
 
-		const reveal = this.props.isRevealingCards ?
+		const revealToggle = this.props.isRevealingCards ?
 			<Tooltipper title="Hide cards" desktop="%s (X)">
 				<button onClick={this.props.onClickHideCards}><i className="joust-fa joust-fa-eye-slash"></i></button>
 			</Tooltipper> :
 			<Tooltipper title="Reveal cards" desktop="%s (X)">
-				<button onClick={this.props.onClickRevealCards} disabled={!this.props.canRevealCards}><i className="joust-fa joust-fa-eye"></i></button>
+				<button onClick={this.props.onClickRevealCards} disabled={!this.props.canRevealCards}>
+					<i className="joust-fa joust-fa-eye"></i>
+				</button>
 			</Tooltipper>;
 
-		const swap = <Tooltipper title="Swap players" desktop="%s (C)">
-			<button onClick={this.props.swapPlayers} disabled={!this.state.canInteract}><i className="joust-fa joust-fa-unsorted"></i></button>
+		const swapToggle = <Tooltipper title="Swap players" desktop="%s (C)">
+			<button onClick={this.props.swapPlayers} disabled={!this.state.canInteract}><i
+				className="joust-fa joust-fa-unsorted"></i></button>
 		</Tooltipper>;
 
 		const log = <Tooltipper title={this.props.isLogVisible ? "Hide event log" : "Show event log"}>
-			<button onClick={this.props.toggleLog} disabled={!this.state.canInteract}><i className="joust-fa joust-fa-bars"></i></button>
+			<button onClick={this.props.toggleLog} disabled={!this.state.canInteract}><i
+				className="joust-fa joust-fa-bars"></i></button>
 		</Tooltipper>;
+
+		const settingsToggleButton = <button
+			onClick={() => this.setState({isShowingSettings: !this.state.isShowingSettings})}>
+			<i className="joust-fa joust-fa-cog"></i>
+		</button>;
+
+		const settingsToggle = !this.state.isShowingSettings ? <Tooltipper title={"Settings"}>
+			{settingsToggleButton}
+		</Tooltipper> : settingsToggleButton;
 
 		return (
 			<div className="joust-scrubber">
 				{this.state.canRewind && !this.state.canPlay ? restart : playpause}
-				<Tooltipper title="Playback speed" desktop="%s (W/S)"><SpeedSelector speed={this.state.speed}
-							   speeds={Scrubber.SPEEDS}
-							   selectSpeed={(speed: number) => this.selectSpeed(speed)}
-							   disabled={!this.state.canInteract}
-				/></Tooltipper>
-				<Timeline duration={this.state.duration }
+				{speedSelector}
+				<Timeline
+					duration={this.state.duration }
 					at={this.state.currentTime }
 					seek={this.props.scrubber.seek.bind(this.props.scrubber) }
 					turnMap={this.props.scrubber.getHistory().turnMap}
 					swapPlayers={this.props.isSwapped}
 					disabled={!this.state.canInteract}
 					ref={(inhibitor) => this.props.scrubber.setInhibitor(inhibitor) }
-					/>
-				{reveal}
-				{swap}
-				{log}
-				{fullscreen}
+				/>
+				{revealToggle}
+				{swapToggle}
+				{this.state.isShowingSettings && <Settings
+					locale={this.props.locale}
+					onSelectLocale={this.props.onSelectLocale && ((locale: string, loaded: () => void) => this.selectLocale(locale, loaded))}
+					isLogVisible={this.props.isLogVisible}
+					onToggleLog={() => this.props.toggleLog()}
+					onClose={() => this.setState({isShowingSettings: false})}
+				/>}
+				{settingsToggle}
+				{fullscreenToggle}
 			</div>
 		);
 	}
 
-	protected play (): void {
+	protected play(): void {
 		this.props.scrubber.play();
 	}
 
@@ -303,5 +348,13 @@ export default class Scrubber extends React.Component<ScrubberProps, ScrubberSta
 			path: "/",
 		});
 		this.props.scrubber.setSpeed(speed);
+	}
+
+	protected selectLocale(locale: string, loaded: () => void): void {
+		cookie.set("joust_locale", locale, {
+			expires: 365, // one year
+			path: "/",
+		});
+		this.props.onSelectLocale(locale, loaded);
 	}
 }